+++ packages/fighting-design/popup/src/props.ts
@@ -6,8 +6,7 @@ import {
   setFunctionProp
 } from '../../_utils'
 import type { ExtractPropTypes } from 'vue'
-import type { PopupDirection } from './interface'
-import type { HandleEvent } from '../../_interface'
+import type { PopupDirection, PopupCallback } from './interface'
 
 export const Props = {
   /** 绑定值，控制是否展示 */
@@ -47,13 +46,13 @@ export const Props = {
   /** 自定义内边距 */
   padding: setStringNumberProp(),
   /** 打开动画开始的回调 */
-  onOpen: setFunctionProp<HandleEvent>(),
+  onOpen: setFunctionProp<PopupCallback>(),
   /** 关闭动画开始的回调 */
-  onClose: setFunctionProp<HandleEvent>(),
+  onClose: setFunctionProp<PopupCallback>(),
   /** 打开动画结束的回调 */
-  onOpenEnd: setFunctionProp<HandleEvent>(),
+  onOpenEnd: setFunctionProp<PopupCallback>(),
   /** 关闭动画结束的回调 */
-  onCloseEnd: setFunctionProp<HandleEvent>()
+  onCloseEnd: setFunctionProp<PopupCallback>()
 } as const
 
 /** popup 组件 props 类型 */